+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Renamed `CMAESState` to `CMAES`.
 - Added `Plot::len`, `Plot::is_empty`, and `Plot::capacity` methods to retrieve the number of data points stored and allocated for.
 - Added `Scale` wrapper type for objective functions that scales the search space.
+- Added `MaxFunctionEvals` and `MaxGenerations` termination criteria.
+- Adjusted the signatures of `CMAES::run` and `CMAES::print_final_info` to reflect the new termination criteria.
 
 ## 0.1.1 (March 4th, 2022)